+++ techTaskmetr/internal/data/response/task_response.go
@@ -6,7 +6,7 @@ import (
 )
 
 type TaskResponse struct {
-	ID          int64     `json:"id"`          // Уникальный идентификатор задачи
+	ID          int       `json:"id"`          // Уникальный идентификатор задачи
 	Title       string    `json:"title"`       // Название задачи
 	Description string    `json:"description"` // Описание задачи
 	CreatedAt   time.Time `json:"created_at"`  // Дата и время создания задачи
