+++ techTaskmetr/internal/repository/repository.go
@@ -11,7 +11,7 @@ type Task interface {
 	Create(ctx *gin.Context, m *models.Task) error
 	Update(ctx *gin.Context, m *models.Task) error
 	Delete(ctx *gin.Context, m *models.Task) error
-	List(ctx *gin.Context, user_id int64) ([]models.Task, error)
+	List(ctx *gin.Context, user_id int) ([]models.Task, error)
 	GetByTitle(ctx *gin.Context, title string, user_id int64) (models.Task, error)
 	GetByID(ctx *gin.Context, id string, user_id int64) (models.Task, error)
 }
